finalize and test application process

+++ app/Http/Controllers/Secretariat/UserController.php
@@ -6,6 +6,7 @@ use App\Http\Controllers\Controller;
 use App\Models\Faculty;
 use App\Models\Role;
 use App\Models\Semester;
+use App\Models\StudyLine;
 use App\Models\User;
 use App\Models\Workshop;
 use App\Models\WorkshopBalance;
@@ -107,9 +108,6 @@ class UserController extends Controller
             'workshop.*' => 'exists:workshops,id',
             'study_line_index' => 'required|array|min:1',
             'email' => 'required|string|email|max:255',
-            'alfonso_language' => ['nullable', Rule::in(array_keys(config('app.alfonso_languages')))],
-            'alfonso_desired_level' => 'nullable|in:B2,C2',
-            'alfonso_passed_by' => 'nullable|date|before:today'
         ]);
 
         $validator->after(function ($validator) use ($request) {
@@ -133,10 +131,7 @@ class UserController extends Controller
             'year_of_acceptance',
             'high_school',
             'neptun',
-            'email',
-            'alfonso_language',
-            'alfonso_desired_level',
-            'alfonso_passed_by'
+            'email'
         ]);
         DB::transaction(function () use ($user, $request, $educational_data) {
             if (!$user->hasEducationalInformation()) {
@@ -145,19 +140,29 @@ class UserController extends Controller
                 $user->educationalInformation->update($educational_data);
             }
 
-            $user->workshops()->sync($request->input('workshop'));
-            $user->faculties()->sync($request->input('faculties'));
+            $user->load('educationalInformation');
 
-            $user->educationalInformation->studyLines()->delete();
-            foreach($request->input('study_line_index') as $index) {
-                $user->educationalInformation->studyLines()->create([
-                    'name' => $request->input('study_line_name_'.$index),
-                    'type' => $request->input('study_line_level_'.$index),
-                    'start' => $request->input('study_line_start_'.$index),
-                    'end' => $request->input('study_line_end_'.$index, null),
-                ]);
+            if($request->has('workshop')){
+                $user->workshops()->sync($request->input('workshop'));
+                WorkshopBalance::generateBalances(Semester::current()->id);
             }
-            WorkshopBalance::generateBalances(Semester::current()->id);
+
+            if($request->has('faculty')){
+                $user->faculties()->sync($request->input('faculty'));
+            }
+
+            if($request->has('study_line_index')){
+                $user->educationalInformation->studyLines()->delete();
+                foreach($request->input('study_line_index') as $index) {
+                    $user->educationalInformation->studyLines()->create([
+                        'name' => $request->input('study_line_name_'.$index),
+                        'type' => $request->input('study_line_level_'.$index),
+                        'start' => $request->input('study_line_start_'.$index),
+                        'end' => $request->input('study_line_end_'.$index, null),
+                    ]);
+                }
+            }
+
         });
